Move helper methods from TestMatchesIterator into a base class
